Moving the Register.jsx to the auth group.
